+++ node/src/rpc.rs
@@ -7,14 +7,14 @@
 
 use std::sync::Arc;
 
-use kumandra_runtime::{opaque::Block, AccountId, Balance, Index as Nonce};
-
 use sc_client_api::AuxStore;
 pub use sc_rpc::{DenyUnsafe, SubscriptionTaskExecutor};
 use sc_transaction_pool_api::TransactionPool;
 use sp_api::ProvideRuntimeApi;
 use sp_block_builder::BlockBuilder;
 use sp_blockchain::{Error as BlockChainError, HeaderBackend, HeaderMetadata};
+
+use kumandra_primitive::{opaque::Block, AccountId, Balance, Index as Nonce};
 
 /// A type representing all RPC extensions.
 pub type RpcExtension = jsonrpsee::RpcModule<()>;
