+++ node/src/service.rs
@@ -9,9 +9,8 @@ use jsonrpsee::RpcModule;
 
 use forests_client_cli::CollatorOptions;
 // Local Runtime Types
-use kumandra_runtime::{
-	opaque::Block, AccountId, Balance, Hash, Index as Nonce, RuntimeApi,
-};
+use kumandra_runtime::RuntimeApi;
+use kumandra_primitive::{opaque::Block, AccountId, Balance, Hash, Index as Nonce};
 
 // Forests Imports
 use forests_client_consensus_aura::{AuraConsensus, BuildAuraConsensusParams, SlotProportion};
