+++ runtime/Cargo.toml
@@ -18,6 +18,7 @@ smallvec = "1.10.0"
 
 # Local
 pallet-template = { path = "../pallets/template", default-features = false }
+kumandra-primitive = { path = "../primitives", default-features = false }
 
 # Substrate
 frame-benchmarking = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.27" , default-features = false, optional = true}
@@ -122,6 +123,7 @@ std = [
 	"xcm-builder/std",
 	"xcm-executor/std",
 	"xcm/std",
+	"kumandra-primitive/std"
 ]
 
 runtime-benchmarks = [
